Seperated each generated event into seperate files
