+++ src/data/characterGeneration/family/siblings.js
@@ -1,0 +1,14 @@
+import { randomInRange } from '../../../helpers/random';
+import { rollFate } from '../../../helpers/helper';
+
+export const generateSibling = () => ({
+	birthOrder: randomInRange(1, 12) <= 6 ? 'younger' : 'older',
+	gender: randomInRange(1, 12) % 2 ? 'brother' : 'sister',
+	fate: rollFate()
+});
+
+// Generates an array of sibling objects
+export const generateSiblings = () =>
+	randomInRange(1, 6) !== 6
+		? Array.from({ length: randomInRange(1, 12) }, () => generateSibling())
+		: [];
